Rydder bort masse gammal moro og bruker resultlist for å få lik visning av resultater overalt
